Refatorando anima-scroll
Refatoração com class do scroll anima
